Add the possibility to place animation sprites absolutely
Now if the continuation byte of a sprite is $02, this means that the
sprite is to be placed absolutely unlike a continuation byte of $01
which place the sprite relatively to the position of the animation.

This allow to hide unused sprites by putting them out of screen (which
is by nature an absolute position). This may also allow to have some
fixed position animation elements.