Implement input event system
It is the way to have inputs that are not consumed more than once. Avoiding the
avatar to chain jump while holding the jump button for example.

It is compatible with the old input system, so this patch only pout the new
system in place, no state use it for now. It will change.