Make it possible to input moves while in thrown state
It makes thrown state an equivalent of SSB tumbling instead of reeling. Making
it possible to react after being hit.

By itself it changes the game from "every move is a potential 0-death combo" to
"there is no possible combos". It prepares for the implementation of hitstun
that will balance between the two extremes.