Implement hitstun
Each player has a hitstun counter that is decremented every frame. While the
hitstun counter is not zero, player states refuse most inputs allowing only
to tech or to influence mid-air direction. When a player is hit, its hitstun
counter is set to a non-zero value, the greater the knockback, the greater the
hitstun value.