Finalize flip animation on the fly
Sprites can be defined as "foreground". A foreground sprite will be placed on
the first available sprite instead of the last one when the animation is
flipped. Foreground sprite are guaranteed to be shown on top of normal sprites
when flipped.

Note that forground sprites do not behave differently than other sprites when
the animation is not flipped. For them to be in foreground in a non-flipped
animation, they have to be defined before other sprites.