Add a particle system
Particles states are in the first half of the $0500 page, in four blocks of this format:

|            Block's header (4 bytes)            |             Particles list (7 * 4 = 28 bytes)                   |
|------------------------------------------------|-------------------------------------------- --------------------|
|   Byte 0    |    Byte 1   |   Byte 2  | Byte 3 |  Bytes 4, 5  |  Bytes 6, 7  | ... | Bytes 28, 29 | Bytes 30, 31 |
| block_param | Tile number | Tile attr | unused | Particle 0 X | Particle 0 Y | ... | Particle 6 X | Particle 6 Y |

A block can be activated (will be processed) or deactivated (not to be processed) by setting block_param to 1 or 0.
Each block contains 7 particles, a specific particle can be hidden by putting it offscreen.