New AI system
There is two main concepts, actions and selectors.

Actions are a series of controller states leading to a desired move.
Example: double-jump is
  * NO-KEY for 1 frame
  * UP for 10 frames
  * NO-KEY for 1 frame
  * UP for 10 frames

Selectors are subroutines that can set an action as the current action.

Each tick, if there is a current action it is continued. If the last action
terminated, selectors are run until one activate an action.