ai: Add condition to attacks data
Attacks now have a list of constraints that must be verified to select the attack.

There are two defined constraints: "the character is facing left" and "the character is facing right". It
allows to use asymetric attacks only if facing the good direction.

There is room to add more as each attacks has space of two 8bit fields.