network: allow to build variants of the network ROM
Assemble with "-DNETWORK_AI" to have the bot playing as player B. This ROM makes for a bot on the network when run by the player B's emulator.

Assemble with "-DNO_NETWORK" to avoid sending/receiving network message. This ROM allows to test the game code itself.

Note that due to hacks and init routines, "-DNETWORK_AI -DNO_NETWORK" is not equivalent to the non-network ROM.