ingame: bugfix characters are no more stuck on platform corners
Corner pixels where considered collision zone for both, the horizontal and
vertical edges. Typically, on a top corner, players where considered on ground
and against the wall, so running toward the platform was blocked.

Corner pixels are now coliding with none of both edges. This allow to fall (and
be against the wall), or move horizontally (and be on ground).