cleaning: ingame state calls move_player the latest possible
Need is growing for a "temporary velocity" which impacts player's velocity only
on the current frame without being impacted by friction, but still colliding
with walls.

To avoid putting this temporary velocity in game state, it must be reset each
frame, and any code needing to set it must be called before move_player.